dynamic_captchas/requirements.txt
Flask==2.3.3
Pillow==10.0.1
Werkzeug==2.3.7
datasets==2.14.0
kaggle==2.0.0
huggingface_hub==0.15.0
numpy<2
pandas<2.2
pyarrow==12.0.1
